A lot of debugging and debug code added to try to track down source of instability with larger problems. This is same issue that bugged us with DagMC cartesian mesh as well. I think having very small fuel FSRs causes problems. Perhaps due to having small optical path lengths or small exponentials. Could be problem with using floats to tally delta psis in this manner. Also added color output.

+++ compute_cell_fission_rates_kernel.c
@@ -12,11 +12,12 @@ void compute_cell_fission_rates_kernel(Parameters P, SimulationData SD, float * 
   uint64_t flux_idx = (uint64_t) cell * P.n_energy_groups;
   scalar_flux += flux_idx;
 
-  float fission_rate = 0.0;
+  double fission_rate = 0.0;
   for( int energy_group = 0; energy_group < P.n_energy_groups; energy_group++ )
   {
     fission_rate += nu_Sigma_f[energy_group] * scalar_flux[energy_group];
   }
 
   SD.readWriteData.cellData.fission_rate[cell] = fission_rate * P.cell_volume;
+  //SD.readWriteData.cellData.fission_rate[cell] = fission_rate;
 }
